Extract UserCreditChecker to a backend module + use localized utility querying to attempt to make it more context-limited (OMS-362)
